sort by creationDate; also consolidate handleAuthRequestAnswered in DeviceManagementComponent (parent)
