+++ libs/angular/src/auth/device-management/device-management-item-group.component.html
@@ -6,8 +6,8 @@
         bit-item-content
         type="button"
         [attr.tabindex]="device.pendingAuthRequest != null ? 0 : null"
-        (click)="approveOrDenyAuthRequest(device.pendingAuthRequest)"
-        (keydown.enter)="approveOrDenyAuthRequest(device.pendingAuthRequest)"
+        (click)="answerAuthRequest(device.pendingAuthRequest)"
+        (keydown.enter)="answerAuthRequest(device.pendingAuthRequest)"
       >
         <!-- Default Content -->
         <span class="tw-text-base">{{ device.displayName }}</span>
